+++ tools/apiview/parsers/js-api-parser/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.3
+
+- add `SkipDiff: true` for dependency header line
+- set related line id for pre-release tags
+- add `export const` before constants
+
 # 2.0.2
 
 - fix issue where `type` is not treated as keyword.